orion-kafka-bridge/requirements.txt
flask
envyaml
kafka-python
gunicorn==20.1.0
requests
simplejson